Web Scraper/requirements.txt
streamlit==1.32.0
langchain
langchain_ollama
selenium==4.18.1
beautifulsoup4==4.12.3
lxml
html5lib
pandas==2.2.1
python-dotenv==1.0.1
requests==2.31.0
ratelimit==2.2.1